chore: delete the ts plugin extension
