--- extensions/vscode-typescript-plugin/.vscodeignore
@@ -1,2 +1,0 @@
-scripts
-typescript-vue-plugin-forward
