--- extensions/vscode-typescript-plugin/CHANGELOG.md
@@ -1,1 +1,0 @@
-For a full change history, refer to the [CHANGELOG](https://github.com/vuejs/language-tools/blob/master/CHANGELOG.md) on the Volar GitHub site.
