--- extensions/vscode-typescript-plugin/package.json
@@ -1,43 +1,0 @@
-{
-	"private": true,
-	"name": "vscode-typescript-vue-plugin",
-	"version": "1.8.26",
-	"repository": {
-		"type": "git",
-		"url": "https://github.com/vuejs/language-tools.git",
-		"directory": "extensions/vscode-typescript-plugin"
-	},
-	"sponsor": {
-		"url": "https://github.com/sponsors/johnsoncodehk"
-	},
-	"icon": "images/icon.png",
-	"displayName": "TypeScript Vue Plugin (Volar)",
-	"description": "Vue Plugin for TypeScript server",
-	"author": "johnsoncodehk",
-	"publisher": "Vue",
-	"engines": {
-		"vscode": "*"
-	},
-	"contributes": {
-		"typescriptServerPlugins": [
-			{
-				"name": "typescript-vue-plugin-bundle",
-				"enableForWorkspaceTypeScriptVersions": true
-			}
-		]
-	},
-	"scripts": {
-		"vscode:prepublish": "npm run build",
-		"prebuild": "cd ../.. && npm run build",
-		"build": "node scripts/build",
-		"watch": "node scripts/build --watch",
-		"pack": "vsce package",
-		"release": "vsce publish",
-		"release:next": "vsce publish --pre-release"
-	},
-	"devDependencies": {
-		"esbuild": "latest",
-		"typescript-vue-plugin": "1.8.26",
-		"vsce": "latest"
-	}
-}
