+++ extensions/vscode/tsconfig.json
@@ -7,5 +7,6 @@
 	"include": [ "src" ],
 	"references": [
 		{ "path": "../../packages/language-server/tsconfig.json" },
+		{ "path": "../../packages/typescript-plugin/tsconfig.json" },
 	],
 }
