+++ package.json
@@ -2,20 +2,17 @@
 	"private": true,
 	"scripts": {
 		"build": "tsc -b",
-		"watch": "npm run build && (npm run watch:base & npm run watch:vue & npm run watch:ts-plugin)",
+		"watch": "npm run build && (npm run watch:base & npm run watch:vue)",
 		"watch:base": "tsc -b -w",
 		"watch:vue": "cd ./extensions/vscode && npm run watch",
-		"watch:ts-plugin": "cd ./extensions/vscode-typescript-plugin && npm run watch",
 		"prerelease": "npm run build && npm run test",
 		"version:test": "lerna version --exact --force-publish --yes --sync-workspace-lock --no-push --no-git-tag-version",
-		"release": "npm run release:base && npm run release:vue && npm run release:ts-plugin",
+		"release": "npm run release:base && npm run release:vue",
 		"release:base": "lerna publish --exact --force-publish --yes --sync-workspace-lock --no-push",
 		"release:vue": "cd ./extensions/vscode && npm run release",
-		"release:ts-plugin": "cd ./extensions/vscode-typescript-plugin && npm run release",
-		"release:next": "npm run release:next-base && npm run release:next-vue && npm run release:next-ts-plugin",
+		"release:next": "npm run release:next-base && npm run release:next-vue",
 		"release:next-base": "npm run release:base -- --dist-tag next",
 		"release:next-vue": "cd ./extensions/vscode && npm run release:next",
-		"release:next-ts-plugin": "cd ./extensions/vscode-typescript-plugin && npm run release:next",
 		"test": "vitest run",
 		"chrome": "vscode-test-web --browserType=chromium --extensionDevelopmentPath=./extensions/vscode ../volar-starter"
 	},
